Enemy Boat
Added a script for an enemy boat. (All it does is follow you and stop after a while, I still need to add a projectile function to it)

Added boat and cannonball model.

Added NavMeshComponents, a library that is used by the script for moving the boat model and detecting the player.

Added and changed layer location of components such as player, enemy, and ground.