Restart-Gameover Menu & Enemy AI
Completed most of the functionality for the enemy AI (Idling, shooting, following)
Added a HP bar for the boat, currently set to 5 hits.
Added a game over screen and restart button.
Removed the previous boat model and replaced it with a pirate ship model.
Fixed the collision properties of the cannonball.